Collect MIT License functions
