+++ rand_distr/tests/common/mod.rs
@@ -1,0 +1,1 @@
+pub mod spec_func;
